refactor: move UnstyledButton to SharedComponents
