+++ src/components/GenericPagesComponents/Select/Select.tsx
@@ -5,7 +5,7 @@ import { useObjectRef } from "@react-aria/utils";
 import { useSelectState } from "react-stately";
 import { useSelect, useButton, mergeProps, useFocusRing } from "react-aria";
 
-import UnstyledButton from "@/components/UnstyledButton";
+import UnstyledButton from "@/components/SharedComponents/UnstyledButton";
 import {
   InputFocusUnderline,
   RotatedInputLabel,
