+++ src/components/SharedComponents/Button/LoadingButton/LoadingButton.tsx
@@ -4,7 +4,7 @@ import Link from "next/link";
 
 import { Spinner } from "./Spinner";
 
-import UnstyledButton from "@/components/UnstyledButton";
+import UnstyledButton from "@/components/SharedComponents/UnstyledButton";
 import { DoubleButtonBorders } from "@/components/SharedComponents/SpriteSheet/BrushSvgs/ButtonBorders/DoubleButtonBorders";
 import Icon, { IconName } from "@/components/SharedComponents/Icon";
 import ButtonLabel from "@/components/SharedComponents/Button/ButtonLabel";
